Javascript: fix missing end code tag
An end of a code block accidentally got nuked in ab72b8b (as part of a conflict
resolution, it looks like), so it's been looking not-quite-right:
http://cl.ly/image/273C1c2d1m35

Just fixing the syntax, though, has a "bad" example that's identical to the
immediately-previous "good" example.  I'm guessing the first bad example is
supposed to show there's no newline at all on the last line.  To try and make
that more evident, I added a visible newline character on the good example,
similar to the second bad example.